feat(streamer): implement stream pending and stream end
Implement stream pending by calling the waker after a 20ms pause. This
makes possible to have a 60fps stream and avoid 0 bytes read files to be
sent.
Implement stream ending when more than 10mb are read. This allows the
stream to not be infinite and finish properly on the caller side.